api/requirements.txt
alembic==1.14.0
fastapi==0.115.6
psycopg[binary]==3.2.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.1
email-validator==2.2.0
